refactor : client - server 통일
