+++ apis/config/privateApi.ts
@@ -1,4 +1,5 @@
-import { onRequest, onRequestError, onResponse, onResponseError } from './interceptor';
+import { onResponse, onResponseError } from './interceptor';
+import { onRequestClient } from './interceptor.client';
 import { BASE_API_URL } from '@/constants';
 import axios from 'axios';
 
@@ -7,7 +8,7 @@ const privateApi = axios.create({
   withCredentials: true,
 });
 
-privateApi.interceptors.request.use(onRequestError);
+privateApi.interceptors.request.use(onRequestClient);
 
 privateApi.interceptors.response.use(onResponse, onResponseError);
