+++ apis/groups/apis.ts
@@ -1,6 +1,4 @@
-import { privateApi } from '../config/privateApi.server';
-
-export const getExample = () => {};
+import privateApi from '../config/privateApi';
 
 type GroupServerResponse = {
   totalCount: number;
@@ -11,6 +9,6 @@ type GroupServerResponse = {
 };
 
 export const getGroupsServer = async () => {
-  const data = await privateApi<GroupServerResponse>(`/api/v1/groups?size=5&page=5`);
+  const data = await privateApi<GroupServerResponse>('/api/v1/groups?size=5&page=5');
   return data;
 };
